smartgrid_api/requirements.txt
Django>=4.0
openpyxl
pytz
djangorestframework
gunicorn
